+++ source/Nuke.Common/Tools/CoverallsNet/CoverallsNet.Generated.cs
@@ -28,11 +28,11 @@ namespace Nuke.Common.Tools.CoverallsNet
         /// <summary><p>Path to the CoverallsNet executable.</p></summary>
         public static string CoverallsNetPath => ToolPathResolver.GetPackageExecutable("coveralls.net", "csmacnz.Coveralls.exe");
         /// <summary><p>Coveralls uploader for .Net Code coverage of your C# source code. Should work with any code files that get reported with the supported coverage tools, but the primary focus is CSharp.</p></summary>
-        public static IEnumerable<string> CoverallsNet(string arguments, string workingDirectory = null, ProcessSettings processSettings = null)
+        public static IEnumerable<string> CoverallsNet(string arguments, string workingDirectory = null, IReadOnlyDictionary<string, string> environmentVariables = null, int? timeout = null, bool redirectOutput = false, Func<string, string> outputFilter = null)
         {
-            var process = ProcessTasks.StartProcess(CoverallsNetPath, arguments, workingDirectory, processSettings?.EnvironmentVariables, processSettings?.ExecutionTimeout, processSettings?.RedirectOutput ?? true);
+            var process = ProcessTasks.StartProcess(CoverallsNetPath, arguments, workingDirectory, environmentVariables, timeout, redirectOutput, outputFilter);
             process.AssertZeroExitCode();
-            return process.Output.Select(x => x.Text);
+            return process.HasOutput ? process.Output.Select(x => x.Text) : null;
         }
         static partial void PreProcess(CoverallsNetSettings toolSettings);
         static partial void PostProcess(CoverallsNetSettings toolSettings);