+++ source/Nuke.Common/Tools/GitVersion/GitVersion.Generated.cs
@@ -28,11 +28,11 @@ namespace Nuke.Common.Tools.GitVersion
         /// <summary><p>Path to the GitVersion executable.</p></summary>
         public static string GitVersionPath => ToolPathResolver.GetPackageExecutable("GitVersion.CommandLine", "GitVersion.exe");
         /// <summary><p>GitVersion is a tool to help you achieve Semantic Versioning on your project.</p></summary>
-        public static IEnumerable<string> GitVersion(string arguments, string workingDirectory = null, ProcessSettings processSettings = null)
+        public static IEnumerable<string> GitVersion(string arguments, string workingDirectory = null, IReadOnlyDictionary<string, string> environmentVariables = null, int? timeout = null, bool redirectOutput = false, Func<string, string> outputFilter = null)
         {
-            var process = ProcessTasks.StartProcess(GitVersionPath, arguments, workingDirectory, processSettings?.EnvironmentVariables, processSettings?.ExecutionTimeout, processSettings?.RedirectOutput ?? true);
+            var process = ProcessTasks.StartProcess(GitVersionPath, arguments, workingDirectory, environmentVariables, timeout, redirectOutput, outputFilter);
             process.AssertZeroExitCode();
-            return process.Output.Select(x => x.Text);
+            return process.HasOutput ? process.Output.Select(x => x.Text) : null;
         }
         static partial void PreProcess(GitVersionSettings toolSettings);
         static partial void PostProcess(GitVersionSettings toolSettings);